@@ -1,17 +1,18 @@
-/*    This file is part of javavp8decoder.
-
-    javavp8decoder is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    javavp8decoder is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with javavp8decoder.  If not, see <http://www.gnu.org/licenses/>.
+/*
+ * This file is part of javavp8decoder.
+ *
+ * javavp8decoder is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * javavp8decoder is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with javavp8decoder.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 package net.sf.javavp8decoder.imageio;
@@ -23,7 +24,7 @@ import javax.imageio.metadata.IIOMetadataFormatImpl;
 public class WebPMetadataFormat extends IIOMetadataFormatImpl {
 
     /** Create a single instance of this class (singleton pattern) */
-    private static WebPMetadataFormat defaultInstance = new WebPMetadataFormat();
+    private static final WebPMetadataFormat defaultInstance = new WebPMetadataFormat();
 
     /** Return the singleton instance */
     public static WebPMetadataFormat getDefaultInstance() {
@@ -48,6 +49,7 @@ public class WebPMetadataFormat extends IIOMetadataFormatImpl {
     }
 
     /** Check for legal element name */
+    @Override
     public boolean canNodeAppear(String elementName, ImageTypeSpecifier imageType) {
         return elementName.equals("KeywordValuePair");
     }
