@@ -1,3 +1,20 @@
+/*
+ * This file is part of javavp8decoder.
+ *
+ * javavp8decoder is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * javavp8decoder is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with javavp8decoder.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -9,14 +26,10 @@ public class IVF {
     private static final int IVF_FRAME_HDR_SZ = 4 + 8;
     private static final int IVF_FILE_HDR_SZ = 32;
 
-    IVF() {
-        InputStream in;
-        try {
-            in = getClass().getResourceAsStream("bbb.ivf");
+    IVF() throws IOException {
+        try (InputStream in = getClass().getResourceAsStream("bbb.ivf")) {
             System.out.println("IVF");
             readFileHeader(in);
-            @SuppressWarnings("unused")
-            int[] frameData;
 
             @SuppressWarnings("unused")
             VP8Decoder f = new VP8Decoder();
@@ -24,19 +37,16 @@ public class IVF {
             @SuppressWarnings("unused")
             int x = 0;
             while (true) {
-                frameData = getFrame(in);
+                @SuppressWarnings("unused")
+                int[] frameData = getFrame(in);
 
                 //f.decodeFrame(frameData, false);
                 x++;
             }
-
-        } catch (IOException e) {
-            // TODO Auto-generated catch block
-            e.printStackTrace();
         }
     }
 
-    private int[] getFrame(InputStream in) throws IOException {
+    private static int[] getFrame(InputStream in) throws IOException {
         int[] frame;
         int c;
         int frameSize = c = in.read();
@@ -57,15 +67,15 @@ public class IVF {
         return frame;
     }
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws IOException {
         new IVF();
     }
 
-    public String toHex(int c) {
+    public static String toHex(int c) {
         return String.format("%1$#x ", c);
     }
 
-    private boolean readFileHeader(InputStream in) throws IOException {
+    private static boolean readFileHeader(InputStream in) throws IOException {
         int c;
         System.out.print("IVF FILE HEADER: ");
         for (int x = 0; x < IVF_FILE_HDR_SZ; x++) {
