@@ -1,17 +1,18 @@
-/*    This file is part of javavp8decoder.
-
-    javavp8decoder is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    javavp8decoder is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with javavp8decoder.  If not, see <http://www.gnu.org/licenses/>.
+/*
+ * This file is part of javavp8decoder.
+ *
+ * javavp8decoder is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * javavp8decoder is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with javavp8decoder.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 package net.sf.javavp8decoder.tools;
@@ -23,6 +24,7 @@ import java.awt.event.WindowEvent;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.Serial;
 
 import javax.imageio.ImageIO;
 import javax.imageio.spi.IIORegistry;
@@ -37,6 +39,7 @@ public class plugintest extends JFrame {
     /**
      *
      */
+    @Serial
     private static final long serialVersionUID = 1L;
 
     public static void main(String[] args) {
@@ -50,6 +53,7 @@ public class plugintest extends JFrame {
         plugintest app;
         app = new plugintest();
         app.addWindowListener(new WindowAdapter() {
+            @Override
             public void windowClosing(WindowEvent e) {
                 System.exit(0);
             }
@@ -62,9 +66,9 @@ public class plugintest extends JFrame {
 
     private BufferedImage bi;
 
-    private JPanel jp;
+    private final JPanel jp;
 
-    private JScrollPane sp;
+    private final JScrollPane sp;
 
     plugintest() {
 
@@ -84,14 +88,16 @@ public class plugintest extends JFrame {
             /**
              *
              */
+            @Serial
             private static final long serialVersionUID = 1L;
 
-            private BufferedImage bi;
+            private final BufferedImage bi;
 
             public ImagePanel(BufferedImage bi) {
                 this.bi = bi;
             }
 
+            @Override
             public void paintComponent(Graphics g) {
                 super.paintComponent(g);
                 g.drawImage(bi, 0, 0, null);
